feat(money): M2/D3 — exact decimals end-to-end (rust_decimal, NUMERIC, JSON-string)
Monetary/Decimal fields are now exact rather than f64, so sales totals (M5) are correct.

- meshble-core: Value::Decimal(rust_decimal::Decimal) (Float kept for non-exact use); domain
  type-check accepts it on Decimal fields; to_json emits a JSON STRING; ComputeInput gains
  decimal()/sum_decimal() (exact-money aggregate). New unit test: 0.1+0.1+0.1 == 0.3 exactly
  (the f64 path gave 0.30000000000000004).
- meshble-db: bind Decimal (sqlx rust_decimal feature); json_to_value parses Decimal from the
  number's canonical string OR a JSON string (no f64 round-trip); reads NUMERIC columns as Decimal
  (dropped the ::float8 read-cast); row_to_json serializes Decimal as a string to preserve precision.
- meshble-schema: OpenAPI Decimal -> {type:string, format:decimal}.
- meshble-server: the list-filter coercion produces exact Value::Decimal for Decimal fields.
- tests: aggregate/nested/compute computes switched to sum_decimal/decimal; the JSON assertion now
  parses the decimal string.

111 workspace tests green. This closes M2 (D3 was the last item).

+++ crates/meshble-db/tests/compute.rs
@@ -8,7 +8,7 @@ use meshble_core::{
 use meshble_db::Db;
 
 fn compute_subtotal(i: &ComputeInput) -> Value {
-    Value::Float(i.float("qty") * i.float("price"))
+    Value::Decimal(i.decimal("qty") * i.decimal("price"))
 }
 meshble_core::inventory::submit! {
     meshble_core::ComputeRegistration { name: "compute_subtotal", func: compute_subtotal }
